+++ app/src/main/java/chatoy/App.java
@@ -1,251 +1,226 @@
 package chatoy;
 
 import component.allframe.BackgroundPanel;
-import component.allframe.DemoScrollBarUI;
-import component.mainframeleft.LeftBox;
-import democlass.database.Room;
-import utils.PathUtils;
 import utils.ScreenUtils;
 
-import javax.imageio.ImageIO;
 import javax.swing.*;
 import javax.swing.border.BevelBorder;
 import javax.swing.border.Border;
-import javax.swing.border.EtchedBorder;
 import javax.swing.plaf.FontUIResource;
-import javax.swing.tree.DefaultTreeCellRenderer;
 import java.awt.*;
 import java.awt.event.*;
-import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
-import java.util.*;
+import java.util.Enumeration;
 
 public class App {
+
     JFrame theFrame = new JFrame("chatoy");
 
-    final int WIDTH = 820;
-    final int HEIGHT = 560;
-    int dividerLocation = 200;
+    final int WIDTH = 550;
+    final int HEIGHT = 550 * 1584 / 1800; // 背景图片1800*1584
+
+    JTextField accountTextField;
+    JPasswordField passwordTextField;
     Font font;
-    Font unchosenFont = new Font("Cabin Sketch",Font.ITALIC,70);
+    Border etchedBorder = BorderFactory.createEtchedBorder(BevelBorder.RAISED,
+            Color.white, Color.pink);
+    Border lineBorder = BorderFactory.createLineBorder(Color.pink);
 
-    Vector<Room> roomVector = new Vector<>();
-    JTextArea roomDescription = new JTextArea();
 
-    // 分割面板
-    JSplitPane jSplitPane = new JSplitPane();
-    // 左侧区域
-    JPanel leftBackPanel;
-    JScrollPane leftScrollPane;
-    JPanel leftPanel;
-
-    Border bevelBorder = BorderFactory.createBevelBorder(BevelBorder.RAISED, Color.black,
-             Color.black, Color.white, Color.pink);
-    Border etchedBorder = BorderFactory.createEtchedBorder(EtchedBorder.RAISED, Color.white, Color.pink);
+    // 组装视图
     public void init() throws IOException {
-
-        // try {
-        //     String lookAndFeel = "com.sun.java.swing.plaf.windows.WindowsLookAndFeel";
-        //
-        //     UIManager.setLookAndFeel(lookAndFeel);
-        //     // UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
-        // } catch (Exception e) {
-        //     e.printStackTrace();
-        // }
-
         // 全局微软雅黑字体
         font = new Font("微软雅黑",Font.PLAIN,12);
         InitGlobalFont(font);
 
-        // 给窗口设置属性
+        // 设置窗口相关属性
         theFrame.setBounds((ScreenUtils.getScreenWidth()-WIDTH)/2, (ScreenUtils.getScreenHeight()-HEIGHT)/2, WIDTH, HEIGHT);
-        // theFrame.setResizable(false);
+        theFrame.setResizable(false);
+        // logo
         theFrame.setIconImage(new ImageIcon("img/logo.png").getImage());
 
-        // 设置菜单栏
-        // JMenuBar jMenuBar = new JMenuBar();
-        // jMenuBar.setBackground(new Color(5, 5, 5));
-        //
-        // JMenu jMenu = new JMenu("设置");
-        // jMenu.setForeground(Color.white);
-        // JMenuItem changeAccount = new JMenuItem("切换账号");
-        // JMenuItem signOut = new JMenuItem("退出程序");
-        // signOut.addActionListener(new ActionListener() {
-        //     @Override
-        //     public void actionPerformed(ActionEvent e) {
-        //         try {
-        //             new Login().init();
-        //             theFrame.dispose();
-        //         } catch (IOException ex) {
-        //             throw new RuntimeException(ex);
-        //         }
-        //     }
-        // });
-        //
-        // signOut.addActionListener(new ActionListener() {
-        //     @Override
-        //     public void actionPerformed(ActionEvent e) {
-        //         System.exit(0);
-        //     }
-        // });
-        // jMenu.add(changeAccount);
-        // jMenu.add(signOut);
-        // jMenuBar.add(jMenu);
-        //
-        // theFrame.setJMenuBar(jMenuBar);
-
-        // 设置分割面板
-        // 支持连续布局
-        jSplitPane = new JSplitPane();
-        jSplitPane.setContinuousLayout(true);
-        jSplitPane.setDividerLocation(dividerLocation);
-        jSplitPane.setDividerSize(1);
-        jSplitPane.setBackground(Color.black);
-
-        // 设置左侧内容
-        Border etchedBorder = BorderFactory.createEtchedBorder(BevelBorder.RAISED,
-                Color.white, Color.pink);
-        // 左侧滚动条
-        leftPanel = new JPanel();
-        // leftPanel.setBackground(new Color(19, 28, 33));
-        leftPanel.setBackground(Color.pink);
-        leftScrollPane = new JScrollPane(leftPanel,ScrollPaneConstants.VERTICAL_SCROLLBAR_ALWAYS,ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER);
-        leftScrollPane.getVerticalScrollBar().setUI(new DemoScrollBarUI());
-        // leftScrollPane.getVerticalScrollBar().setBackground(Color.blue);
-        // leftScrollPane.getVerticalScrollBar().setUI(new BasicScrollBarUI() {
-        //     @Override
-        //     protected void configureScrollBarColors() {
-        //         this.thumbColor = Color.lightGray;
-        //     }
-        // });
-
-
-
-        // 组装左侧区域
-        leftPanel.add(new LeftBox());
-        jSplitPane.setLeftComponent(new LeftBox());
-/*
-        DefaultMutableTreeNode root0 = new DefaultMutableTreeNode("0");
-        DefaultMutableTreeNode root1 = new DefaultMutableTreeNode("1");
-        DefaultMutableTreeNode root2 = new DefaultMutableTreeNode("2");
-        DefaultMutableTreeNode root3 = new DefaultMutableTreeNode("3");
-        DefaultMutableTreeNode root4 = new DefaultMutableTreeNode("4");
-
-        root0.add(root1);
-        root0.add(root2);
-        root0.add(root3);
-        root0.add(root4);
-
-        Color color = new Color(5,5,5);
-        JTree tree = new JTree(root0);
-        // MyRenderer myRenderer = new MyRenderer();
-        // myRenderer.setBackgroundNonSelectionColor(color);
-        // myRenderer.setBackgroundSelectionColor(new Color(140,140,140));
-        // tree.setCellRenderer(myRenderer);
-
-        tree.setBackground(color);
-        // 设置当前tree默认选中root2
-        tree.setSelectionRow(2);
-        tree.addTreeSelectionListener(new TreeSelectionListener() {
-            // 当条目选中变化后，这个方法会执行
-            @Override
-            public void valueChanged(TreeSelectionEvent e) {
-                // 得到当前选中的节点对象
-                Object lastPathComponent = e.getNewLeadSelectionPath().getLastPathComponent();
-
-                if (root0.equals(lastPathComponent)) {
-                    jSplitPane.setRightComponent(new JLabel("这里进行聊天0..."));
-                    jSplitPane.setDividerLocation(220);
-                } else if (root1.equals(lastPathComponent)) {
-                    jSplitPane.setRightComponent(new JLabel("这里进行聊天1..."));
-                    jSplitPane.setDividerLocation(220);
-                } else if (root2.equals(lastPathComponent)) {
-                    jSplitPane.setRightComponent(new JLabel("这里进行聊天2..."));
-                    jSplitPane.setDividerLocation(220);
-                } else if (root3.equals(lastPathComponent)) {
-                    jSplitPane.setRightComponent(new JLabel("这里进行聊天3..."));
-                    jSplitPane.setDividerLocation(220);
-                } else if (root4.equals(lastPathComponent)) {
-                    jSplitPane.setRightComponent(new JLabel("这里进行聊天4..."));
-                    jSplitPane.setDividerLocation(220);
-                }
-
-            }
-        });
-*/
-        // 设置右边内容
-        // 消息输入区
-        Box bottomBox = Box.createHorizontalBox();
-        JButton sendButton = new JButton("发送");
-        // sendButton.setForeground(Color.white);
-        sendButton.setBackground(Color.pink);
-        JTextField textField = new JTextField("请输入");
-        textField.setBorder(etchedBorder);
-        // textField.setBackground(new Color(38,45,49));
-        textField.setBackground(new Color(30, 30, 30));
-        textField.setForeground(Color.gray);
-        textField.addMouseListener(new MouseAdapter() {
-            @Override
-            public void mouseExited(MouseEvent e) {
-                if (textField.getText().equals("")) {
-                    textField.setText("请输入");
-                    textField.setForeground(Color.gray);
-                }
-            }
-            @Override
-            public void mouseEntered(MouseEvent e) {
-                if (textField.getText().equals("请输入")) {
-                    textField.setText("");
-                    textField.setForeground(Color.white);
-                }
-            }
-        });
-
-        bottomBox.add(textField);
-        bottomBox.add(sendButton);
-
+        // 设置窗口的内容
         BackgroundPanel backgroundPanel = new BackgroundPanel(new ImageIcon("img/LoginBackground.png").getImage());
-        backgroundPanel.setLayout(new BorderLayout());
-        backgroundPanel.add(bottomBox, BorderLayout.SOUTH);
 
-        // 设置默认背景
-        JPanel unchosenPanel = new JPanel();
-        unchosenPanel.setLayout(new BorderLayout());
-        unchosenPanel.setBackground(new Color(38, 45 ,49));
-        // 大title
-        JLabel unchosenLabel = new JLabel("Hi, Chatoy !", SwingConstants.CENTER);
-        unchosenLabel.setForeground(Color.pink);
-        unchosenLabel.setFont(unchosenFont);
+        // 组装登录相关的元素
+        Box holeBox = Box.createVerticalBox();
 
-        unchosenPanel.add(unchosenLabel, BorderLayout.CENTER);
+        // 组装用户名
+        Box accountBox = Box.createHorizontalBox();
+        JLabel accountLabel = new JLabel("账号：");
+        accountLabel.setForeground(Color.pink); // 字体设为粉色
+        accountTextField = new JTextField(15);
+        accountTextField.setText("请输入用户名");
+        accountTextField.setForeground(Color.gray); // 文本框字体设为浅灰色
+        accountTextField.setOpaque(false); // 文本框设置为透明
+        accountTextField.setBorder(lineBorder); // 粉色边框
+
+        accountBox.add(accountLabel);
+        accountBox.add(Box.createHorizontalStrut(6));
+        accountBox.add(accountTextField);
+
+        // 组装密码
+        Box passwordBox = Box.createHorizontalBox();
+        JLabel passwordLabel = new JLabel("密码：");
+        passwordLabel.setForeground(Color.pink); // 字体设为粉色
+        passwordTextField = new JPasswordField(15);
+        passwordTextField.setForeground(Color.white); // 文本框字体设为浅灰色
+        passwordTextField.setOpaque(false); // 文本框设置为透明
+        passwordTextField.setBorder(lineBorder); // 粉色边框
+
+        passwordBox.add(passwordLabel);
+        passwordBox.add(Box.createHorizontalStrut(6));
+        passwordBox.add(passwordTextField);
+
+        // 输入网址
+        Box httpBox = Box.createHorizontalBox();
+        JLabel httpLabel = new JLabel("网址：");
+        httpLabel.setForeground(Color.pink); // 字体设为粉色
+        JTextField httpTextField = new JTextField();
+        httpTextField.setText("请输入网址");
+        httpTextField.setForeground(Color.gray); // 文本框字体设为浅灰色
+        httpTextField.setOpaque(false); // 文本框设置为透明
+        httpTextField.setBorder(etchedBorder); // 粉色边框
+
+        httpBox.add(httpLabel);
+        httpBox.add(Box.createHorizontalStrut(6));
+        httpBox.add(httpTextField);
+
+        // mouseListener
+        passwordTextField.addMouseListener(new MouseAdapter() {
+            @Override
+            public void mouseClicked(MouseEvent e) {
+                super.mouseClicked(e);
+                if (e.getButton() == MouseEvent.BUTTON1) {
+                    if (accountTextField.getText().equals("")) {
+                        accountTextField.setText("请输入用户名");
+                        accountTextField.setForeground(Color.gray);
+                    }
+                    if (httpTextField.getText().equals("")) {
+                        httpTextField.setText("请输入网址");
+                        httpTextField.setForeground(Color.gray);
+                    }
+                }
+            }
+        });
+        accountTextField.addMouseListener(new MouseAdapter() {
+            @Override
+            public void mouseClicked(MouseEvent e) {
+                super.mouseClicked(e);
+                if (e.getButton() == MouseEvent.BUTTON1) {
+                    if (accountTextField.getText().equals("请输入用户名")) {
+                        accountTextField.setText("");
+                        accountTextField.setForeground(Color.white);
+                    }
+                    if (httpTextField.getText().equals("")) {
+                        httpTextField.setText("请输入网址");
+                        httpTextField.setForeground(Color.gray);
+                    }
+                }
+            }
+        });
+        httpTextField.addMouseListener(new MouseAdapter() {
+            @Override
+            public void mouseClicked(MouseEvent e) {
+                super.mouseClicked(e);
+                if (e.getButton() == MouseEvent.BUTTON1) {
+                    if (accountTextField.getText().equals("")) {
+                        accountTextField.setText("请输入用户名");
+                        accountTextField.setForeground(Color.gray);
+                    }
+                    if (httpTextField.getText().equals("请输入网址")) {
+                        httpTextField.setText("");
+                        httpTextField.setForeground(Color.white);
+                    }
+                }
+            }
+        });
+
+        // 组装按钮
+        Box buttonBox = Box.createHorizontalBox();
+        JButton loginButton = new JButton("登录");
+        loginButton.setBackground(Color.pink);
+        JButton registerButton = new JButton("注册");
+        registerButton.setBackground(Color.pink);
+
+        loginButton.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                // 获取用户输入的数据
+                String account = accountTextField.getText();
+                String password = passwordTextField.getPassword().toString();
+
+                // 直接登录现在
+                try {
+                    new Chatoy().init();
+                    theFrame.dispose();
+                } catch (IOException ex) {
+                    throw new RuntimeException(ex);
+                }
 
 
-        jSplitPane.setRightComponent(unchosenPanel);
+                // 访问登录接口
 
-        theFrame.add(jSplitPane);
+            }
+        });
 
-        // setVisible
+        registerButton.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                // 跳转到注册页面
+                try {
+                    new Register().init();
+                } catch (IOException ex) {
+                    throw new RuntimeException(ex);
+                }
+                // 当前界面消失
+                theFrame.dispose();
+            }
+        });
+
+        buttonBox.add(loginButton);
+        buttonBox.add(Box.createHorizontalStrut(50));
+        buttonBox.add(registerButton);
+
+        // Chatoy大title
+        JLabel title = new JLabel("chatoy", SwingConstants.CENTER);
+        Font titleFont=new Font("Cabin Sketch",Font.ITALIC,70);
+        title.setForeground(Color.pink);
+        title.setFont(titleFont);
+        title.setLocation(50,50);
+        Box titleBox = Box.createHorizontalBox();
+        titleBox.add(title);
+
+        holeBox.add(Box.createVerticalStrut(20));
+        holeBox.add(titleBox);
+        holeBox.add(Box.createVerticalStrut(25));
+        holeBox.add(accountBox);
+        holeBox.add(Box.createVerticalStrut(15));
+        holeBox.add(passwordBox);
+        holeBox.add(Box.createVerticalStrut(15));
+        holeBox.add(httpBox);
+        holeBox.add(Box.createVerticalStrut(20));
+        holeBox.add(buttonBox);
+
+        // 创建一个纯色底面
+
+        JPanel jPanel = new JPanel();
+        jPanel.setBorder(etchedBorder);
+        jPanel.setBounds(105,65,340,310);
+        jPanel.setBackground(new Color(5, 5, 5));
+        jPanel.add(holeBox);
+        backgroundPanel.setLayout(null);
+        // backgroundPanel.setLayout(new GridBagLayout());
+        backgroundPanel.add(jPanel);
+
         theFrame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        theFrame.add(backgroundPanel);
         theFrame.setVisible(true);
     }
 
     public static void main(String[] args) throws IOException {
-        App chatoy = new App();
-        chatoy.init();
-
-
-
-        // chatoy.addRoom(new Room("ccc",333, "this 1"));
-        // // 添加聊天室
-        // chatoy.addRoom(new Room("test1111111111111111111",333, "this 1"));
-        // chatoy.addRoom(new Room("tes333333333333333333333333333333333333333333333333333333333333333333333333t",12, "this 2"));
-        // chatoy.addRoom(new Room("testromm1",333, "this 1"));
-        // chatoy.addRoom(new Room("testromm1",333, "this 1"));
-        // chatoy.addRoom(new Room("testromm1",333, "this 1"));
-        // chatoy.addRoom(new Room("testromm1",333, "this 1"));
-        // chatoy.addRoom(new Room("testromm1",333, "this 1"));
-        // chatoy.addRoom(new Room("testromm1",333, "this 1"));
+        App login = new App();
+        login.init();
     }
 
     private static void InitGlobalFont(Font font) {
@@ -257,46 +232,6 @@ public class App {
             if (value instanceof FontUIResource) {
                 UIManager.put(key, fontRes);
             }
-        }
-    }
-
-    // 自定义节点绘制器
-    private class MyRenderer extends DefaultTreeCellRenderer {
-        private ImageIcon root0Icon = null;
-        private ImageIcon root1Icon = null;
-        private ImageIcon root2Icon = null;
-        private ImageIcon root3Icon = null;
-        private ImageIcon root4Icon = null;
-
-        public MyRenderer() {
-            root0Icon = new ImageIcon(PathUtils.getRealPath("logo.png"));
-            root1Icon = new ImageIcon(PathUtils.getRealPath("logo.png"));
-            root2Icon = new ImageIcon(PathUtils.getRealPath("logo.png"));
-            root3Icon = new ImageIcon(PathUtils.getRealPath("logo.png"));
-            root4Icon = new ImageIcon(PathUtils.getRealPath("logo.png"));
-        }
-
-        // 当绘制树的每个节点时，都会调用这个方法
-        @Override
-        public Component getTreeCellRendererComponent(JTree tree, Object value, boolean sel, boolean expanded, boolean leaf, int row, boolean hasFocus) {
-            super.getTreeCellRendererComponent(tree, value, sel, expanded, leaf, row, hasFocus);
-
-            ImageIcon image = null;
-            switch (row) {
-                case 0:
-                    image = root0Icon;
-                case 1:
-                    image = root1Icon;
-                case 2:
-                    image = root2Icon;
-                case 3:
-                    image = root3Icon;
-                case 4:
-                    image = root4Icon;
-            }
-
-            this.setIcon(image);
-            return this;
         }
     }
 }
